- Refactor TupleIndexHashTableIteratorTest
