+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/LeftTupleIndexHashTableIteratorTest.java
@@ -15,15 +15,9 @@
 package org.drools.mvel;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 
 import org.drools.core.RuleBaseConfiguration;
-import org.drools.core.base.ClassFieldAccessorCache;
-import org.drools.core.base.ClassFieldAccessorStore;
-import org.drools.core.base.ClassObjectType;
-import org.drools.core.base.evaluators.EvaluatorRegistry;
-import org.drools.core.base.evaluators.Operator;
 import org.drools.core.common.BetaConstraints;
 import org.drools.core.common.InternalFactHandle;
 import org.drools.core.common.SingleBetaConstraints;
@@ -31,22 +25,14 @@ import org.drools.core.impl.KnowledgeBaseFactory;
 import org.drools.core.reteoo.BetaMemory;
 import org.drools.core.reteoo.LeftTupleImpl;
 import org.drools.core.reteoo.NodeTypeEnums;
-import org.drools.core.rule.Declaration;
-import org.drools.core.rule.Pattern;
 import org.drools.core.spi.BetaNodeFieldConstraint;
-import org.drools.core.spi.InternalReadAccessor;
 import org.drools.core.util.AbstractHashTable;
 import org.drools.core.util.Entry;
 import org.drools.core.util.Iterator;
 import org.drools.core.util.index.TupleIndexHashTable;
 import org.drools.core.util.index.TupleIndexHashTable.FieldIndexHashTableFullIterator;
 import org.drools.core.util.index.TupleList;
-import org.drools.model.functions.Predicate2;
-import org.drools.model.index.BetaIndexImpl;
-import org.drools.modelcompiler.util.EvaluationUtil;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.kie.api.KieBase;
 import org.kie.api.runtime.KieSession;
 
@@ -57,23 +43,10 @@ import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-@RunWith(Parameterized.class)
-public class LeftLeftTupleIndexHashTableIteratorTest {
+public class LeftTupleIndexHashTableIteratorTest extends BaseTupleIndexHashTableIteratorTest {
 
-    public static EvaluatorRegistry registry = new EvaluatorRegistry();
-
-    private final boolean useLambdaConstraint;
-
-    public LeftLeftTupleIndexHashTableIteratorTest(boolean useLambdaConstraint) {
+    public LeftTupleIndexHashTableIteratorTest(boolean useLambdaConstraint) {
         this.useLambdaConstraint = useLambdaConstraint;
-    }
-
-    @Parameterized.Parameters(name = "useLambdaConstraint={0}")
-    public static Collection<Object[]> getParameters() {
-        Collection<Object[]> parameters = new ArrayList<>();
-        parameters.add(new Object[]{false});
-        parameters.add(new Object[]{true});
-        return parameters;
     }
 
     @Test
@@ -129,75 +102,32 @@ public class LeftLeftTupleIndexHashTableIteratorTest {
         leftTupleList.add(new LeftTupleImpl(fh13, null, true));
         ((TupleList) hashTable.getTable()[0]).setNext(leftTupleList);
 
-        Entry[] table = hashTable.getTable();
-        List list = new ArrayList();
-        for (int i = 0; i < table.length; i++) {
-            if (table[i] != null) {
-                List entries = new ArrayList();
-                entries.add(i);
-                Entry entry = table[i];
-                while (entry != null) {
-                    entries.add(entry);
-                    entry = entry.getNext();
-                }
-                list.add(entries.toArray());
-            }
-        }
-        assertEquals(5, list.size());
+        List tableIndexList = createTableIndexListForAssertion(hashTable);
+        assertEquals(5, tableIndexList.size());
 
         // This tests the hashcode index allocation. If the rehash function (or any other way hashcodes are computed) changes, these numbers will change.
         // standard-drl and exec-model have different getRightExtractor().getIndex() value so hashCode changes
-
         if (useLambdaConstraint) {
-            Object[] entries = (Object[]) list.get(0);
-            assertEquals(0, entries[0]);
-            assertEquals(3, entries.length);
-
-            entries = (Object[]) list.get(1);
-            assertEquals(49, entries[0]);
-            assertEquals(3, entries.length);
-
-            entries = (Object[]) list.get(2);
-            assertEquals(51, entries[0]);
-            assertEquals(3, entries.length);
-
-            entries = (Object[]) list.get(3);
-            assertEquals(60, entries[0]);
-            assertEquals(2, entries.length);
-
-            entries = (Object[]) list.get(4);
-            assertEquals(61, entries[0]);
-            assertEquals(2, entries.length);
+            assertTableIndex(tableIndexList, 0, 0, 3);
+            assertTableIndex(tableIndexList, 1, 49, 3);
+            assertTableIndex(tableIndexList, 2, 51, 3);
+            assertTableIndex(tableIndexList, 3, 60, 2);
+            assertTableIndex(tableIndexList, 4, 61, 2);
         } else {
-            Object[] entries = (Object[]) list.get(0);
-            assertEquals(0, entries[0]);
-            assertEquals(3, entries.length);
-
-            entries = (Object[]) list.get(1);
-            assertEquals(102, entries[0]);
-            assertEquals(2, entries.length);
-
-            entries = (Object[]) list.get(2);
-            assertEquals(103, entries[0]);
-            assertEquals(2, entries.length);
-
-            entries = (Object[]) list.get(3);
-            assertEquals(115, entries[0]);
-            assertEquals(3, entries.length);
-
-            entries = (Object[]) list.get(4);
-            assertEquals(117, entries[0]);
-            assertEquals(3, entries.length);
+            assertTableIndex(tableIndexList, 0, 0, 3);
+            assertTableIndex(tableIndexList, 1, 102, 2);
+            assertTableIndex(tableIndexList, 2, 103, 2);
+            assertTableIndex(tableIndexList, 3, 115, 3);
+            assertTableIndex(tableIndexList, 4, 117, 3);
         }
 
-        list = new ArrayList<LeftTupleImpl>();
+        List resultList = new ArrayList<LeftTupleImpl>();
         Iterator it = betaMemory.getLeftTupleMemory().iterator();
         for (LeftTupleImpl leftTuple = (LeftTupleImpl) it.next(); leftTuple != null; leftTuple = (LeftTupleImpl) it.next()) {
-            list.add(leftTuple);
+            resultList.add(leftTuple);
         }
 
-        assertEquals(13, list.size());
-
+        assertEquals(13, resultList.size());
     }
 
     @Test
@@ -238,71 +168,5 @@ public class LeftLeftTupleIndexHashTableIteratorTest {
                    sameInstance((Object) tuples[2]));
         assertThat(iterator.next(),
                    is((Object) null));
-
     }
-
-    private static BetaNodeFieldConstraint createFooThisEqualsDBetaConstraint(boolean useLambdaConstraint) {
-        if (useLambdaConstraint) {
-            Pattern pattern = new Pattern(0, new ClassObjectType(Foo.class));
-            Pattern varPattern = new Pattern(1, new ClassObjectType(Foo.class));
-            Predicate2<Foo, Foo> predicate = new Predicate2.Impl<Foo, Foo>((_this, d) -> EvaluationUtil.areNullSafeEquals(_this, d));
-            BetaIndexImpl<Foo, Foo, Foo> index = new BetaIndexImpl<Foo, Foo, Foo>(Foo.class, org.drools.model.Index.ConstraintType.EQUAL, 1, _this -> _this, d -> d, Foo.class);
-            return LambdaConstraintTestUtil.createLambdaConstraint2(Foo.class, Foo.class, pattern, varPattern, "d", predicate, index);
-        } else {
-            ClassFieldAccessorStore store = new ClassFieldAccessorStore();
-            store.setClassFieldAccessorCache(new ClassFieldAccessorCache(Thread.currentThread().getContextClassLoader()));
-            store.setEagerWire(true);
-            InternalReadAccessor extractor = store.getReader(Foo.class,
-                                                             "this");
-            Declaration declaration = new Declaration("d",
-                                                      extractor,
-                                                      new Pattern(0,
-                                                                  new ClassObjectType(Foo.class)));
-
-            String expression = "this " + Operator.EQUAL.getOperatorString() + " d";
-            return new MVELConstraintTestUtil(expression, declaration, extractor);
-        }
-    }
-
-    public static class Foo {
-
-        private String val;
-        private int hashCode;
-
-        public Foo(String val,
-                   int hashCode) {
-            this.val = val;
-            this.hashCode = hashCode;
-        }
-
-        public String getVal() {
-            return val;
-        }
-
-        @Override
-        public int hashCode() {
-            return hashCode;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj)
-                return true;
-            if (obj == null)
-                return false;
-            if (getClass() != obj.getClass())
-                return false;
-            Foo other = (Foo) obj;
-            if (hashCode != other.hashCode)
-                return false;
-            if (val == null) {
-                if (other.val != null)
-                    return false;
-            } else if (!val.equals(other.val))
-                return false;
-            return true;
-        }
-
-    }
-
 }
